WF5 — Cloud Platforms conformance (6 crates) + review fixes (#244)
* docs(wf5): Cloud Platforms workflow spec + plan (6C/19W/18I)

Workflow 5 of the conformance roadmap: armature-{aws,azure,gcp,lambda,
cloudrun,azure-functions}. 43 findings across the six cloud crates —
unwired enable_* services, ignored credential sources, config knobs read
nowhere, unpopulated runtime data, non-compiling READMEs, and crates with
no tests. Implement-vs-remove decided per finding against the pinned SDK
version. Six edit-only agents on disjoint crates; coordinator runs the
central gate + bookkeeping.


* feat(wf5): Cloud Platforms conformance — 43 findings across 6 crates

Workflow 5 of the conformance roadmap: armature-{aws,azure,gcp,lambda,
cloudrun,azure-functions}. All 43 findings closed (6 Critical, 19 Warning,
18 Info), 73 new regression tests. Implement-vs-remove decided per finding
against each crate's pinned SDK version.

Unwired services made real (or removed):
- azure: real ServiceBusClient (SAS model, azure_messaging_servicebus 0.21)
  + servicebus()/cosmos_database() accessors (2 Critical).
- gcp: secret-manager/cloud-run/cloud-functions as authenticated REST
  clients; firestore REMOVED (no client in google-cloud-googleapis 0.16)
  (3 Critical).

Credential sources honored:
- aws: Environment/IamRole/WebIdentity now map to distinct real providers
  (were collapsed into the default chain).
- gcp: config.credentials threaded into every client builder (1 Critical).
- azure: ConnectionString/StorageAccountKey removed (SDK supports neither;
  azure_identity 1.0 dropped EnvironmentCredential).

Config knobs wired / runtime data populated:
- lambda: Cognito authorizer claims (V2 jwt.claims / V1 map), path_parameters,
  stage_variables now populated (were hardcoded empty); impl_request_handler!
  forwards the full request (1 Critical).
- azure-functions: max_request_size (413) + timeout_seconds (504) enforced.
- cloudrun: real health-check HTTP endpoint; service_url from CLOUD_RUN_URL;
  metadata::fetch surfaces unreachable-server errors; concurrent fetch.
- gcp: emulator_host + service_configs endpoint overrides.

Non-compiling READMEs rewritten across all six crates; azure-functions'
inert bindings module and aws' placeholder `pub mod s3` deleted.

Breaking (0.x -> minor): azure, gcp, aws, cloudrun -> 0.2.0; additive:
lambda, azure-functions -> 0.1.3. Internal dependents repinned.

Gate: fmt; workspace clippy (full-with-saml -D warnings); per-crate
clippy --all-features and --no-default-features; per-crate test
--all-features (73 pass); workspace test (Docker-required); cargo audit.
These crates stay outside the MSRV-1.89 gate and default CI matrix
(above-1.89 vendor SDKs; pre-existing, flagged in CHANGELOG). TODO totals
395 -> 87; only WF9 (Core & Tooling) remains.


* chore: raise workspace MSRV from 1.89 to 1.94.1

The Cloud Platforms crates (WF5) pull vendor SDKs whose own MSRV exceeds
1.89 — aws-smithy-runtime-api / aws-smithy-types / aws-smithy-schema /
aws-types all require rustc 1.94.1 — so the workspace could not honestly
claim 1.89 with those members present.

- root Cargo.toml: [workspace.package] and umbrella [package] rust-version
  1.89 -> 1.94.1
- CI: MSRV Check job -> 1.94.1 toolchain, renamed "MSRV Check (1.94.1)"
- armature-mcp: drop the bespoke 1.91 override (it existed only to clear
  the old 1.89 floor; the workspace floor is now higher) -> inherit workspace
- AGENTS.md + CHANGELOG updated

Verified on 1.94.1: umbrella `--features full-with-saml` and every cloud
crate `--all-features` build clean.

BREAKING CHANGE: minimum supported Rust is now 1.94.1.


* fix(wf5): code-review pass — timeouts, OOM guard, versioning, CI coverage

Addresses all 30 findings from a 7-dimension review of the Cloud Platforms
branch (2 Critical, 6 High, 8 Medium, 14 Low).

Critical:
- gcp: REST clients (Secret Manager/Cloud Run/Cloud Functions) now build with
  request + connect timeouts — were untimed, so a stalled endpoint hung the
  calling task forever.
- lambda: bump 0.1.3 -> 0.2.0; impl_request_handler! changed the required
  handle_request signature (a breaking macro contract shipped as a patch).

High:
- azure-functions: max_request_size enforced during ingestion via a bounded
  Limited body (a chunked upload with no Content-Length could OOM before the
  check fired); request timeout now also covers ingestion (408); removed the
  dead InputBinding/OutputBinding scaffolding; bump 0.1.3 -> 0.2.0.
- aws: deleted 11 hollow placeholder modules (dynamodb..cognito) left behind
  after the s3 removal.
- ci: added all six cloud crates to the test-members matrix (--all-features)
  so their feature-gated conformance suites actually run.

Medium/Low: gcp status-specific error mapping (+3 GcpError variants), cloudrun
accept-loop resilience + metadata 5xx surfacing, azure ServiceBus timeout +
connection-string key redaction, gcp percent-encoded path segments +
https-default endpoints, aws from_env credential precedence aligned to the AWS
SDK chain, internal dedup (gcp RestClient/rest_service!, aws get_or_init!), and
new boundary/URL regression tests. CHANGELOG and TODO reconciled.

Also (pre-existing, unblocks the pre-commit hook under stable 1.97.1's clippy):
trivial redundant-reference removals in armature-i18n and an unnecessary i64
cast in armature-jwt.

Gate (stable 1.97.1): fmt; clippy --workspace --all-targets
--features full-with-saml -D warnings; 81 cloud-crate tests green;
--no-default-features compile; armature-storage consumer build; rustls-only
verified; cargo audit clean.

BREAKING CHANGE: armature-lambda and armature-azure-functions are raised to
0.2.0 — the lambda impl_request_handler! macro contract changed, and
azure-functions removed the InputBinding/OutputBinding traits.


---------

+++ armature-jwt/src/service.rs
@@ -302,7 +302,7 @@ mod tests {
         // (refresh_token.exp - now ≈ refresh_expires_in, within tolerance of ±2 seconds)
         let now = chrono::Utc::now().timestamp();
         let actual_lifetime = r_exp - now;
-        let expected = pair.refresh_expires_in as i64;
+        let expected = pair.refresh_expires_in;
         assert!(
             (actual_lifetime - expected).abs() <= 2,
             "TokenPair.refresh_expires_in {} must match refresh token lifetime {} (±2s)",
